fix formatting warning found by Xcode analyze

+++ src/command/game/LoadGameCommand.m
@@ -375,7 +375,7 @@ static const int maxStepsForReplayMoves = 10;
 
       if (actualNumberOfHandicapStones != expectedNumberOfHandicapStones)
       {
-        *errorMessage = [NSString stringWithFormat:@"The handicap (%ld) is greater than the number of black setup stones (%ld).", expectedNumberOfHandicapStones, actualNumberOfHandicapStones];
+        *errorMessage = [NSString stringWithFormat:@"The handicap (%ld) is greater than the number of black setup stones (%lu).", expectedNumberOfHandicapStones, (unsigned long)actualNumberOfHandicapStones];
         return false;
       }
     }
